Update check and email for FileController

+++ src/main/java/com/ooad/xproject/controller/FileController.java
@@ -6,6 +6,7 @@ import com.ooad.xproject.bo.StudentImportBO;
 import com.ooad.xproject.bo.SvResult;
 import com.ooad.xproject.config.FileConfig;
 import com.ooad.xproject.constant.RespStatus;
+import com.ooad.xproject.dto.StudentDTO;
 import com.ooad.xproject.entity.*;
 import com.ooad.xproject.mapper.ProjectMapper;
 import com.ooad.xproject.mapper.RecordInstMapper;
@@ -23,8 +24,13 @@ import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.File;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.ooad.xproject.constant.ProjInstStatus.Confirm;
 
 
 // todo: role + database -> directory name -> upload/download
@@ -54,7 +60,10 @@ public class FileController {
 
     private final ResourceMapper resourceMapper;
 
-    public FileController(FileConfig fileConfig, FileService fileService, ExcelService excelService, StudentService studentService, RoleService roleService, TeacherService teacherService, RecordService recordService, RecordInstMapper recordInstMapper, ProjectMapper projectMapper, ProjInstService projInstService, SubmissionInstService submissionInstService, PermissionService permissionService, SubmissionMapper submissionMapper, ResourceMapper resourceMapper) {
+    private final ProjectService projService;
+    private final MailService mailService;
+
+    public FileController(FileConfig fileConfig, FileService fileService, ExcelService excelService, StudentService studentService, RoleService roleService, TeacherService teacherService, RecordService recordService, RecordInstMapper recordInstMapper, ProjectMapper projectMapper, ProjInstService projInstService, SubmissionInstService submissionInstService, PermissionService permissionService, SubmissionMapper submissionMapper, ResourceMapper resourceMapper, ProjectService projService, MailService mailService) {
         this.fileConfig = fileConfig;
         this.fileService = fileService;
         this.excelService = excelService;
@@ -69,6 +78,8 @@ public class FileController {
         this.permissionService = permissionService;
         this.submissionMapper = submissionMapper;
         this.resourceMapper = resourceMapper;
+        this.projService = projService;
+        this.mailService = mailService;
     }
 
     @PostMapping("api/upload")
@@ -141,6 +152,11 @@ public class FileController {
         MultipartFile[] files = uploadFileProjIdVO.getFiles();
         int projId = uploadFileProjIdVO.getProjId();
 
+        // check project accessible
+        if (!projService.isAccessible(projId)) {
+            return new Result<>(RespStatus.FAIL, "Project is not accessible");
+        }
+
         String filePath = fileService.upload(files[0], fileConfig.getInputRoot(), "input.xlsx");
 //        System.out.println(filePath);
         List<RecordUnitBO> recordUnitBOList = excelService.readRecordUnitBO(filePath);
@@ -172,6 +188,12 @@ public class FileController {
     public Result<Integer> postProjStdExcel(@RequestBody UploadFileProjIdVO uploadFileProjIdVO) {
         MultipartFile[] files = uploadFileProjIdVO.getFiles();
         int projId = uploadFileProjIdVO.getProjId();
+
+        // check project accessible
+        if (!projService.isAccessible(projId)) {
+            return new Result<>(RespStatus.FAIL, "Project is not accessible");
+        }
+
         String filePath = fileService.upload(files[0], fileConfig.getInputRoot(), "input.xlsx");
 //        System.out.println(filePath);
         List<StudentClassBO> studentClassBOList = excelService.readStudentClassBO(filePath);
@@ -208,7 +230,18 @@ public class FileController {
 
         Submission submission = submissionMapper.selectByPrimaryKey(sbmId);
 
+        // check project accessible
+        if (!projService.isAccessible(submission.getProjId())) {
+            return new Result<>(RespStatus.FAIL, "Project is not accessible");
+        }
+
         ProjectInst projectInst = projInstService.getPIByProjIdAndStdRoleId(submission.getProjId(), role.getRoleId());
+
+        if (projectInst == null) {
+            return new Result<>(RespStatus.FAIL, "No team yet");
+        } else if (!projectInst.getStatus().equals(Confirm.name())) {
+            return new Result<>(RespStatus.FAIL, "Team is not confirm");
+        }
 
         SubmissionInst submissionInst = new SubmissionInst();
         submissionInst.setSbmId(sbmId);
@@ -216,9 +249,22 @@ public class FileController {
         submissionInst.setSubmitterId(role.getRoleId());
         submissionInst.setAttachments(attachment.toString());
         if (submissionInstService.upsertSubmissionInst(submissionInst) == 0) {
-            return new Result<>(RespStatus.FAIL);
+            return new Result<>(RespStatus.FAIL, "No chance left");
         }
         File studentDir = fileService.getOrCreateStudentDir(submissionInst);
+
+        // send email to all students
+        SimpleDateFormat formatter= new SimpleDateFormat("yyyy-MM-dd 'at' HH:mm:ss z");
+        Date date = new Date(System.currentTimeMillis());
+
+        List<StudentDTO> stdList = projInstService.getStudentDTOByProjInstId(projectInst.getProjInstId());
+        List<String> mailList = stdList.stream().map(StudentDTO::getEmail).collect(Collectors.toList());
+        mailService.sendMailToStudent(mailList, "[XProject] Your team has submitted an assignment",
+                "Your team has submitted an assignment\r\n" +
+                        "Title: " + submission.getTitle() + "\r\n" +
+                        "Submitter: " + role.getUsername() + "\r\n" +
+                        "Submission time: " + formatter.format(date) + "\r\n" +
+                        "This automatic notification message was sent by Xproject");
 
         fileService.deleteFilesExceptFolder(studentDir);
         return new Result<>(fileService.upload(files, studentDir.getPath()));
@@ -228,6 +274,14 @@ public class FileController {
     public ResponseEntity<byte[]> getAllSbmFiles(HttpServletRequest request, @RequestParam("sbmId") int sbmId
             , @RequestHeader("user-agent") String userAgent
             , @RequestParam(required = false, defaultValue = "false") boolean inline) {
+
+        Submission submission = submissionMapper.selectByPrimaryKey(sbmId);
+
+        // check project accessible
+        if (!projService.isAccessible(submission.getProjId())) {
+            // TODO LYZ
+            return null;
+        }
 
         File file = fileService.getSbmDir(sbmId);
         String outputPath = fileConfig.getOutputRoot() + "\\" + "output.zip";
@@ -239,6 +293,11 @@ public class FileController {
     public Result<?> postResources(@RequestBody UploadFileProjIdVO uploadFileProjIdVO) {
         MultipartFile[] files = uploadFileProjIdVO.getFiles();
         int projId = uploadFileProjIdVO.getProjId();
+
+        // check project accessible
+        if (!projService.isAccessible(projId)) {
+            return new Result<>(RespStatus.FAIL, "Project is not accessible");
+        }
 
         Role role = roleService.getByUsername(RoleUtils.getUsername());
         int creatorId = role.getRoleId();
@@ -260,8 +319,20 @@ public class FileController {
             , @RequestHeader("user-agent") String userAgent
             , @RequestParam(required = false, defaultValue = "false") boolean inline) {
 
-        File file = fileService.getResDir(srcId);
         Resource resource = resourceMapper.selectByPrimaryKey(srcId);
+
+        if (resource == null) {
+            // TODO LYZ
+            return null;
+        }
+
+        // check project accessible
+        if (!projService.isAccessible(resource.getProjId())) {
+            // TODO LYZ
+            return null;
+        }
+
+        File file = fileService.getResDir(srcId);
         String targetPath = file.getPath() + "\\" + resource.getFileName();
         return fileService.download(request, targetPath, userAgent, resource.getFileName(), inline);
     }
@@ -269,6 +340,11 @@ public class FileController {
     @ResponseBody
     @GetMapping("api/all/resource/list")
     public Result<?> getResourceList(@RequestParam("projId") int projId) {
+
+        // check project accessible
+        if (!projService.isAccessible(projId)) {
+            return new Result<>(RespStatus.FAIL, "Project is not accessible");
+        }
 
         List<Resource> resources = resourceMapper.selectByProjId(projId);
 
@@ -286,8 +362,14 @@ public class FileController {
     @GetMapping("api/teacher/project/resource/del")
     public Result<?> getDeleteResources(@RequestParam("srcId") int srcId) {
         Resource resource = resourceMapper.selectByPrimaryKey(srcId);
+
         if (resource == null) {
-            return new Result<>(RespStatus.FAIL);
+            return new Result<>(RespStatus.FAIL, "No such resource");
+        }
+
+        // check project accessible
+        if (!projService.isAccessible(resource.getProjId())) {
+            return new Result<>(RespStatus.FAIL, "Project is not accessible");
         }
         File file = fileService.getResDir(srcId);
         fileService.deleteFilesAndFolder(file);
